+++ src/app/enums/list-actions.model.ts
@@ -2,5 +2,6 @@ export enum ListActions {
     deleteList,
     modifyList,
     selectList,
-    completeList
+    completeList,
+    addList
 }